requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
reportlab
